// app/acteurs/types.ts
export interface ActorType {
  id: number
  actor: string
  actor_sigle: string
  email: string
  phone: string
  whatsapp: string
  actor_type_id: number
  headquarter_photo?: string
  logo?: string
  address: string
  latitude?: number
  longitude?: number
  password?: string
  code: string
  description?: string
}